+++ babel.config.js
@@ -5,38 +5,17 @@ const plugins = [
         'loose': true
     }],
     '@babel/transform-literals',
-    '@babel/transform-function-name',
-    '@babel/transform-arrow-functions',
     '@babel/transform-block-scoped-functions',
     ['@babel/plugin-transform-classes', {
         'loose': true
     }],
-    '@babel/transform-object-super',
-    '@babel/transform-shorthand-properties',
-    ['@babel/transform-computed-properties', {
-        'loose': true
-    }],
-    ['@babel/transform-for-of', {
-        'loose': true
-    }],
     '@babel/transform-sticky-regex',
     '@babel/transform-unicode-regex',
-    ['@babel/transform-spread', {
-        'loose': true
-    }],
-    '@babel/transform-parameters', ['@babel/transform-destructuring', {
-        'loose': true
-    }],
     '@babel/transform-block-scoping',
-    '@babel/plugin-transform-member-expression-literals',
-    '@babel/transform-property-literals',
-    '@babel/transform-async-to-generator',
-    '@babel/transform-regenerator',
     ['@babel/transform-runtime', {
         'regenerator': true
     }],
-    '@babel/proposal-class-properties',
-    '@babel/proposal-object-rest-spread'
+    '@babel/proposal-class-properties'
 ];
 
 let presets = [
@@ -58,11 +37,10 @@ if (process.env['NODE_ENV'] === 'es5') {
             {
                 loose: true,
                 targets: {
-                    edge: '17',
-                    firefox: '60',
-                    chrome: '67',
-                    safari: '11.1',
-                    ie: '11'
+                    edge: '107',
+                    firefox: '107',
+                    chrome: '108',
+                    safari: '16.2'
                 },
                 useBuiltIns: false
             }]